pic1

+++ src/com/alensw/Jandan/PicActivity1.java
@@ -1,5 +1,9 @@
 package com.alensw.Jandan;
 
+import android.annotation.SuppressLint;
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.graphics.Point;
 import android.net.Uri;
 import android.os.AsyncTask;
 import android.os.Bundle;
@@ -7,15 +11,18 @@ import android.support.v4.view.PagerAdapter;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.ActionBarActivity;
 import android.util.Log;
+import android.view.Display;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.ImageView;
+import android.widget.ProgressBar;
+import android.widget.TextView;
 import com.alensw.gif.GifImageView;
+import com.alensw.ui.PictureView;
 import com.alensw.ui.TouchImageView;
 
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
+import java.io.*;
+import java.lang.ref.WeakReference;
 import java.net.URL;
 import java.net.URLConnection;
 import java.util.ArrayList;
@@ -38,10 +45,10 @@ public class PicActivity1 extends ActionBarActivity {
 		boolean isGif;
 	}
 
-	;
-	private ArrayList<ItemInfo> mItems;
+	private ArrayList<ItemInfo> mItems = new ArrayList<>();
 	private int progress = 0;
 	private GifLoader mGifLoader;
+	private PictureLoader mPictureLoader;
 
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
@@ -107,6 +114,7 @@ public class PicActivity1 extends ActionBarActivity {
 		@Override
 		public void setPrimaryItem(ViewGroup container, int position, Object object) {
 			if (mPicAdapter != object) {
+				mPicAdapter = (PicAdapter) object;
 				mPicAdapter.start();
 			}
 		}
@@ -126,14 +134,208 @@ public class PicActivity1 extends ActionBarActivity {
 				mGifImageView = (GifImageView) mView;
 				progress = 0;
 				mGifLoader = new GifLoader();
-				mGifLoader.execute(null, null);
+				mGifLoader.execute(mItemInfo.url, null);
 			} else {
-
+				mTouchImageView = (TouchImageView) mView;
+				BitmapLoader bl = new BitmapLoader(mTouchImageView, null, null);
+				bl.execute(mItemInfo.url);
 			}
 		}
 	}
 
-	;
+	private class BitmapLoader extends AsyncTask<String, Void, Bitmap> {
+
+		/**
+		 * Weak reference to the target {@link android.widget.ImageView} where the bitmap will be loaded into.
+		 *
+		 * Using a weak reference will avoid memory leaks if the target ImageView is retired from memory before the load finishes.
+		 */
+		private final WeakReference<PictureView> mImageViewRef;
+
+		/**
+		 * Weak reference to the target {@link android.widget.TextView} where error messages will be written.
+		 *
+		 * Using a weak reference will avoid memory leaks if the target ImageView is retired from memory before the load finishes.
+		 */
+		private final WeakReference<TextView> mMessageViewRef;
+
+
+		private final WeakReference<ProgressBar> mProgressWheelRef;
+
+
+		/**
+		 * Error message to show when a load fails
+		 */
+		private int mErrorMessageId;
+
+
+		/**
+		 * Constructor.
+		 *
+		 * @param imageView     Target {@link android.widget.ImageView} where the bitmap will be loaded into.
+		 */
+		public BitmapLoader(PictureView imageView, TextView messageView, ProgressBar progressWheel) {
+			mImageViewRef = new WeakReference<>(imageView);
+			mMessageViewRef = new WeakReference<>(messageView);
+			mProgressWheelRef = new WeakReference<>(progressWheel);
+		}
+
+
+		@Override
+		protected Bitmap doInBackground(String... params) {
+			Bitmap result = null;
+			if (params.length != 1) return result;
+			String hashName = Utilities.md5(params[0]);
+			FileCache.mCacheDir.mkdirs();
+			File picture = FileCache.generateCacheFile(hashName);
+			try {
+				if (picture != null) {
+					//Decode file into a bitmap in real size for being able to make zoom on the image
+					result = BitmapFactory.decodeStream(new FlushedInputStream
+							(new BufferedInputStream(new FileInputStream(picture))));
+				}
+				if (result == null) {
+
+				}
+			} catch (OutOfMemoryError e) {
+				// If out of memory error when loading image, try to load it scaled
+				result = loadScaledImage(picture.getPath());
+				if (result == null) {
+
+				}
+			} catch (NoSuchFieldError e) {
+			} catch (Throwable t) {
+			}
+			return result;
+		}
+
+		@Override
+		protected void onPostExecute(Bitmap result) {
+			hideProgressWheel();
+			if (result != null) {
+				showLoadedImage(result);
+			} else {
+				showErrorMessage();
+			}
+		}
+
+		@SuppressLint("InlinedApi")
+		private void showLoadedImage(Bitmap result) {
+			if (mImageViewRef != null) {
+				final PictureView imageView = mImageViewRef.get();
+				if (imageView != null) {
+					imageView.setBitmap(result);
+					imageView.setImageBitmap(result);
+					imageView.setVisibility(View.VISIBLE);
+					// mBitmap  = result;
+				} // else , silently finish, the fragment was destroyed
+			}
+			if (mMessageViewRef != null) {
+				final TextView messageView = mMessageViewRef.get();
+				if (messageView != null) {
+					messageView.setVisibility(View.GONE);
+				} // else , silently finish, the fragment was destroyed
+			}
+		}
+
+		private void showErrorMessage() {
+			if (mImageViewRef != null) {
+				final ImageView imageView = mImageViewRef.get();
+				if (imageView != null) {
+					// shows the default error icon
+					imageView.setVisibility(View.VISIBLE);
+				} // else , silently finish, the fragment was destroyed
+			}
+			if (mMessageViewRef != null) {
+				final TextView messageView = mMessageViewRef.get();
+				if (messageView != null) {
+					messageView.setText(mErrorMessageId);
+					messageView.setVisibility(View.VISIBLE);
+				} // else , silently finish, the fragment was destroyed
+			}
+		}
+
+		private void hideProgressWheel() {
+			if (mProgressWheelRef != null) {
+				final ProgressBar progressWheel = mProgressWheelRef.get();
+				if (progressWheel != null) {
+					progressWheel.setVisibility(View.GONE);
+				}
+			}
+		}
+
+	}
+
+	static class FlushedInputStream extends FilterInputStream {
+		public FlushedInputStream(InputStream inputStream) {
+			super(inputStream);
+		}
+
+		@Override
+		public long skip(long n) throws IOException {
+			long totalBytesSkipped = 0L;
+			while (totalBytesSkipped < n) {
+				long bytesSkipped = in.skip(n - totalBytesSkipped);
+				if (bytesSkipped == 0L) {
+					int byteValue = read();
+					if (byteValue < 0) {
+						break;  // we reached EOF
+					} else {
+						bytesSkipped = 1; // we read one byte
+					}
+				}
+				totalBytesSkipped += bytesSkipped;
+			}
+			return totalBytesSkipped;
+		}
+	}
+
+	private Bitmap loadScaledImage(String storagePath) {
+		// set desired options that will affect the size of the bitmap
+		BitmapFactory.Options options = new BitmapFactory.Options();
+		options.inScaled = true;
+		options.inPurgeable = true;
+		if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.GINGERBREAD_MR1) {
+			options.inPreferQualityOverSpeed = false;
+		}
+		if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.HONEYCOMB) {
+			options.inMutable = false;
+		}
+		// make a false load of the bitmap - just to be able to read outWidth, outHeight and outMimeType
+		options.inJustDecodeBounds = true;
+		BitmapFactory.decodeFile(storagePath, options);
+
+		int width = options.outWidth;
+		int height = options.outHeight;
+		int scale = 1;
+
+		Display display = getWindowManager().getDefaultDisplay();
+		Point size = new Point();
+		int screenWidth;
+		int screenHeight;
+		if (android.os.Build.VERSION.SDK_INT >= android.os.Build.VERSION_CODES.HONEYCOMB_MR2) {
+			display.getSize(size);
+			screenWidth = size.x;
+			screenHeight = size.y;
+		} else {
+			screenWidth = display.getWidth();
+			screenHeight = display.getHeight();
+		}
+
+		if (width > screenWidth) {
+			// second try to scale down the image , this time depending upon the screen size
+			scale = (int) Math.floor((float)width / screenWidth);
+		}
+		if (height > screenHeight) {
+			scale = Math.max(scale, (int) Math.floor((float)height / screenHeight));
+		}
+		options.inSampleSize = scale;
+
+		// really load the bitmap
+		options.inJustDecodeBounds = false; // the next decodeFile call will be real
+		return BitmapFactory.decodeFile(storagePath, options);
+
+	}
 
 	private class GifLoader extends AsyncTask<String, Integer, Uri> {
 		private Long mLength;
@@ -141,7 +343,12 @@ public class PicActivity1 extends ActionBarActivity {
 		protected Uri doInBackground(String... params) {
 			InputStream is;
 			try {
-				File file = FileCache.generateCacheFile(params[1]);
+				String hashName = Utilities.md5(params[0]) + ".gif";
+				FileCache.mCacheDir.mkdirs();
+				File file = FileCache.generateCacheFile(hashName);
+				if (file.exists())
+					return Uri.fromFile(file);
+
 				FileOutputStream os = new FileOutputStream(file);
 
 				URL url = new URL(params[0]);
